+++ CHANGES.md
@@ -9,6 +9,7 @@ Release Notes.
 - Remove Bloom filter for dictionary-encoded tags.
 - Implement BanyanDB MCP.
 - Support deleting non-entity tags when updating the schema.
+- Remove check requiring tags in criteria to be present in projection.
 
 ### Bug Fixes